requirements.txt
numpy==1.21.0
pandas==0.25.3
Pillow==8.2.0
Pillow==9.5.0
streamlit==0.62.0
